stop using extern and macros

+++ src/gfx/MainLayer.h
@@ -35,7 +35,7 @@ private:
     FileBrowser m_Browser;
     std::chrono::high_resolution_clock::time_point now = std::chrono::high_resolution_clock::now();
     std::chrono::high_resolution_clock::time_point last = now;
-    int m_ZoomLevel{1};
+    float m_ZoomLevel{1.0f};
     bool m_bFileLoaded{false};
     std::string m_RomName;
 };
